fix: set email to username after create/get
The username retrieved from Cognito is a UUID and not an email, this
mismatch between Cognito and Profile users causes issues when deleting
users.
Set the email in to the username field to fix the delete scenario.
Note that this will still fail if the email address is modified and no
longer matches the profile username.

TIS21-2149