feat: build JSON endpoints
This creates endpoints that are as ReSTful as other TIS APIs.
Exploring this as an alternative to managing OAuth requests.

In-house framework methods enrich the principal.
That relies on a token being passed in the request.

This reuses the service a pattern for quicker testing and validation.

This does not resolve all 'ReST' issues.
For example, all managed exceptions yield a "Bad Request" response.

TIS21-6616: Validate PoC OAuth handling
TIS21-6687: Angular UI for User Management